Revert __all__ related changes from #82
