+++ array_api_compat/_internal.py
@@ -5,7 +5,6 @@ Internal helpers
 from functools import wraps
 from inspect import signature
 
-
 def get_xp(xp):
     """
     Decorator to automatically replace xp with the corresponding array module.
@@ -45,31 +44,3 @@ specification for more details.
         return wrapped_f
 
     return inner
-
-
-def _get_all_public_members(module, exclude=None, extend_all=False):
-    """Get all public members of a module.
-
-    Parameters
-    ----------
-    module : module
-        The module to get members from.
-    exclude : callable, optional
-        A callable that takes a name and returns True if the name should be
-        excluded from the list of members.
-    extend_all : bool, optional
-        If True, extend the module's __all__ attribute with the members of the
-        module derived from dir(module). To be used for libraries that do not have a complete __all__ list.
-    """
-    members = getattr(module, "__all__", [])
-
-    if members and not extend_all:
-        return members
-
-    if exclude is None:
-        exclude = lambda name: name.startswith("_")  # noqa: E731
-
-    members = members + [_ for _ in dir(module) if not exclude(_)]
-
-    # remove duplicates
-    return list(set(members))
