+++ array_api_compat/common/_typing.py
