+++ array_api_compat/cupy/_aliases.py
@@ -2,16 +2,15 @@ from __future__ import annotations
 
 from functools import partial
 
-import cupy as cp
-
 from ..common import _aliases
-from ..common import _linalg
 
 from .._internal import get_xp
 
 asarray = asarray_cupy = partial(_aliases._asarray, namespace='cupy')
 asarray.__doc__ = _aliases._asarray.__doc__
+del partial
 
+import cupy as cp
 bool = cp.bool_
 
 # Basic renames
@@ -74,28 +73,7 @@ if hasattr(cp, 'isdtype'):
 else:
     isdtype = get_xp(cp)(_aliases.isdtype)
 
-
-cross = get_xp(cp)(_linalg.cross)
-outer = get_xp(cp)(_linalg.outer)
-EighResult = _linalg.EighResult
-QRResult = _linalg.QRResult
-SlogdetResult = _linalg.SlogdetResult
-SVDResult = _linalg.SVDResult
-eigh = get_xp(cp)(_linalg.eigh)
-qr = get_xp(cp)(_linalg.qr)
-slogdet = get_xp(cp)(_linalg.slogdet)
-svd = get_xp(cp)(_linalg.svd)
-cholesky = get_xp(cp)(_linalg.cholesky)
-matrix_rank = get_xp(cp)(_linalg.matrix_rank)
-pinv = get_xp(cp)(_linalg.pinv)
-matrix_norm = get_xp(cp)(_linalg.matrix_norm)
-svdvals = get_xp(cp)(_linalg.svdvals)
-diagonal = get_xp(cp)(_linalg.diagonal)
-trace = get_xp(cp)(_linalg.trace)
-
-# These functions are completely new here. If the library already has them
-# (i.e., numpy 2.0), use the library version instead of our wrapper.
-if hasattr(cp.linalg, 'vector_norm'):
-    vector_norm = cp.linalg.vector_norm
-else:
-    vector_norm = get_xp(cp)(_linalg.vector_norm)
+__all__ = _aliases.__all__ + ['asarray', 'asarray_cupy', 'bool', 'acos',
+                              'acosh', 'asin', 'asinh', 'atan', 'atan2',
+                              'atanh', 'bitwise_left_shift', 'bitwise_invert',
+                              'bitwise_right_shift', 'concat', 'pow']
