+++ array_api_compat/dask/array/_aliases.py
@@ -1,18 +1,42 @@
 from __future__ import annotations
 
-from functools import partial
-from typing import TYPE_CHECKING
-
-import numpy as np
-
-from ..._internal import get_xp
-from ...common import _aliases, _linalg
+from ...common import _aliases
 from ...common._helpers import _check_device
 
-if TYPE_CHECKING:
-    from typing import Optional, Tuple, Union
+from ..._internal import get_xp
 
-    from ...common._typing import Device, Dtype, ndarray
+import numpy as np
+from numpy import (
+    # Constants
+    e,
+    inf,
+    nan,
+    pi,
+    newaxis,
+    # Dtypes
+    bool_ as bool,
+    float32,
+    float64,
+    int8,
+    int16,
+    int32,
+    int64,
+    uint8,
+    uint16,
+    uint32,
+    uint64,
+    complex64,
+    complex128,
+    iinfo,
+    finfo,
+    can_cast,
+    result_type,
+)
+
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from typing import Optional, Union
+    from ...common._typing import ndarray, Device, Dtype
 
 import dask.array as da
 
@@ -24,7 +48,6 @@ astype = _aliases.astype
 # This arange func is modified from the common one to
 # not pass stop/step as keyword arguments, which will cause
 # an error with dask
-
 
 # TODO: delete the xp stuff, it shouldn't be necessary
 def dask_arange(
@@ -49,11 +72,11 @@ def dask_arange(
     args.append(step)
     return xp.arange(*args, dtype=dtype, **kwargs)
 
-
 arange = get_xp(da)(dask_arange)
 eye = get_xp(da)(_aliases.eye)
 
-asarray = partial(_aliases._asarray, namespace="dask.array")
+from functools import partial
+asarray = partial(_aliases._asarray, namespace='dask.array')
 asarray.__doc__ = _aliases._asarray.__doc__
 
 linspace = get_xp(da)(_aliases.linspace)
@@ -89,22 +112,34 @@ trunc = get_xp(np)(_aliases.trunc)
 matmul = get_xp(np)(_aliases.matmul)
 tensordot = get_xp(np)(_aliases.tensordot)
 
+from dask.array import (
+    # Element wise aliases
+    arccos as acos,
+    arccosh as acosh,
+    arcsin as asin,
+    arcsinh as asinh,
+    arctan as atan,
+    arctan2 as atan2,
+    arctanh as atanh,
+    left_shift as bitwise_left_shift,
+    right_shift as bitwise_right_shift,
+    invert as bitwise_invert,
+    power as pow,
+    # Other
+    concatenate as concat,
+)
 
-EighResult = _linalg.EighResult
-QRResult = _linalg.QRResult
-SlogdetResult = _linalg.SlogdetResult
-SVDResult = _linalg.SVDResult
-qr = get_xp(da)(_linalg.qr)
-cholesky = get_xp(da)(_linalg.cholesky)
-matrix_rank = get_xp(da)(_linalg.matrix_rank)
-matrix_norm = get_xp(da)(_linalg.matrix_norm)
+# exclude these from all since
+_da_unsupported = ['sort', 'argsort']
 
+common_aliases = [alias for alias in _aliases.__all__ if alias not in _da_unsupported]
 
-def svdvals(x: ndarray) -> Union[ndarray, Tuple[ndarray, ...]]:
-    # TODO: can't avoid computing U or V for dask
-    _, s, _ = da.linalg.svd(x)
-    return s
+__all__ = common_aliases + ['asarray', 'bool', 'acos',
+                            'acosh', 'asin', 'asinh', 'atan', 'atan2',
+                            'atanh', 'bitwise_left_shift', 'bitwise_invert',
+                            'bitwise_right_shift', 'concat', 'pow',
+                            'e', 'inf', 'nan', 'pi', 'newaxis', 'float32', 'float64', 'int8',
+                            'int16', 'int32', 'int64', 'uint8', 'uint16', 'uint32', 'uint64',
+                            'complex64', 'complex128', 'iinfo', 'finfo', 'can_cast', 'result_type']
 
-
-vector_norm = get_xp(da)(_linalg.vector_norm)
-diagonal = get_xp(da)(_linalg.diagonal)
+del da, partial, common_aliases, _da_unsupported,
